pyproject.toml
[tool.poetry]
name = "fast-api-learn"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
openai = "^1.30.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
